Revert "Scene"
This reverts commit 2018d7f102c2e9f18fb69685b2b86729f33695d7.

+++ Editor_Window/Engine_SourceCode/tTime.cpp
@@ -32,7 +32,7 @@ namespace t
 
 		wchar_t str[50] = L"";
 		swprintf_s(str, 50, L"Time : %d", (int)fps);
-		int len = wcsnlen_s(str, (size_t)50);
+		int len = wcsnlen_s(str, 50);
 
 		TextOut(hdc, 0, 0, str, len);
 	}
